Limit the spurious interrupt messages
Under some emulators, the kernel shows an excessive amount of messages on
spurious interrupt detection. These messages can be really annoying and they
difficult to have a successfull session with the system.

To avoid them, it has been introduced the constant MAX_SPU_NOTICES (with a
default value of 10) to limit the number of these messages. Once the number
of messages is bigger than this value they no longer appear in the console.