Fixed a PANIC on systems with 16GB+ of RAM
A misplaced bitwise operator AND was not ensuring the limitation to 1GB as
the amount of physical memory supported. As a consequence of that, on systems
with 16GB or more of RAM (perhaps even on 8GB, not tested) the kernel paniced
due to an overflow in the 4MB temporary Page Directory that is setup during
the kernel start up. #2