Added VMA_REGIONS as a general configurable option (default: 1000)
From now on the vma array in every process has no longer a fixed size
of PAGE_SIZE / sizeof(struct vma), instead its size is now defined in
the 'VMA_REGIONS' option in config.h which defaults to 1000.

The vma struct has a size of 32 bytes, so until now, the vma array in
every process could include only up to 128 entries, which was clearly
not enough when porting (compiling) large programs like Perl.