Finally 'bg_used_dirs_count' is honored in the Ext2 filesystem
Added support to update the 'bg_used_dirs_count' value from the
group descriptors. Such value needs to be updated on every directory
creation or deletion.

This forced to change the filesystem semantics in all the inode
allocation functions to include a second argument that specifies the
type of the file that is going to be created. Unfortunately this new
argument is only useful in the Ext2 filesystem.