Saved twice the 'signum' value before calling the signal handler
I've experienced kernel panics after returning from a call to a signal
handler because the 'signum' value was clobbered by the printf() function.

The workaround has been to push twice the %eax register before calling
the signal handler and discard the extra push on return.

Is this a bug in Newlib or in my side?