Added support for the framebuffer console (fbcon)
This is a (hopefully) complete framebuffer console which introduces a
uniform handling of the virtual consoles under the i386 platforms,
with a true bitmapped graphical consoles.

The switch from VGA text mode to graphical mode is made by GRUB after
loading the kernel. Since FiwixOS is using GRUB 0.97 (aka GRUB legacy),
it has been needed to patch GRUB in order to include the 'vbeset' and
'vbematch' commands that use the VBE functions to do all the job.

This patched version of GRUB will be included by default in the next
FiwixOS 4 release.

This is experimental and still lacks some features.
More commits to come ...