Added support for interrupt nesting
This is a major change in how the kernel works internally. Nested
interrupts permit the benefits of a true multitasking by keeping
the clock tick always on either during exceptions and system
calls.

Among others, this has the benefit that is now possible to
implement a set of SysRq key combinations(like in Linux), that
should help to know the kernel state by performing various
low-level commands.