Reorganized the signal functions
This will avoid that signals sent by the kernel have to use the same
checking mechanism as if they were sent by a user process.

This also fixes a permission problem when sending the SIGCHLD signal
to a parent process owned by a different user, during the context of
sys_exit().