Added the configuration option CONFIG_SYSCALL_6TH_ARG
This is intended to add more Linux 2.0 ABI compatibility to support the
original sys_ipc() system call, which requires up to 6 arguments.

Fiwix by default uses a maximum of 5 arguments per system call, so any
binary built on FiwixOS will use a different implementation of the
sys_ipc(), and also any other system call that requires more than 5
arguments. This breaks compatibility with Linux 2.0 ABI.

In order to include the 6th argument, Fiwix will use the register 'ebp'
which might lead to some problems. Use this with caution.

This option should only be used if you need to execute a Linux 2.0
binary and, for some reason, you cannot rebuild it on FiwixOS.