Added DMA support to ATA disk driver
This support is actually for Multi-word DMA mode only, and for now, it
only works on the Intel PCI 82371SB PIIX3 IDE controller, which is the
same one that comes with Bochs and QEMU.

Since these emulators make no difference in speed or performance between
all DMA transfer modes (Single-word DMA, Multi-word DMA and Ultra-DMA),
the result is the same for any of DMA mechanisms. The point here is that
the kernel no longer needs to copy word-by-word the buffer contents using
I/O ports (which is really expensive), instead, the controller reads and
writes directly on the buffer. This clearly results in disk transfers of
several orders of magnitude faster.