Now kernel mounts the root filesystem in rw mode by default
This change introduces the new kernel parameter 'ro' to force the kernel
to mount the root filesystem in read-only mode.

The next version of FiwixOS will take care of this. Meanwhile, if you
plan to upgrade the kernel of your FiwixOS system, you'll need to add
the parameter 'ro' to the GRUB line in order to keep the same behavior.

E.g.:
kernel /boot/fiwix ro root=/dev/hda2