Added support for UNIX domain sockets  #36
This should create the foundations to help to add networking support
into Fiwix. The code is heavily influenced by the Linux kernel, mainly
because I try to keep Linux 2.0 i386 ABI compatibility.

It still lacks the system calls sendto() and recvfrom() so datagrams
are not possible yet, but the work is in progress.