Improved the efficiency of the buffer cache
Until now the buffer cache was using a whole memory page (4096 bytes)
for every single buffer, regardless of the real size of the buffer
(1024, 2048 or 4096 bytes), leading to a major internal fragmentation
and a lot of memory not used.

From now on, the buffer cache is defined with three lists of buffer
heads for free buffers; corresponding to buffer sizes of 1024, 2048
and 4096 bytes. This means that for a 1KB block size filesystem there
is now 4 times more buffers in cache than before. So it's expected a
reduction of disk accesses which will improve the system performance
significantly.

This change was needed to prepare the kernel to include the ability
to make multiple block device requests in one shot, which should
improve dramatically the disk access time.