Rewritten how multiple I/O block requests are managed
Until now, when the kernel needed to fill a memory page with the contents
of a file, (e.g.  during the demand loading of an executable, or when
reading any other file), it read every block of the file in a single
operation, and didn't start reading the next block until the previous
one finished. This led to constant context switching (sleep() and wakeup()
calls) that degraded the system performance; for instance, on a filesystem
with a block size of 1KiB this means to read or write 4 blocks per memory
page, which means 4 context switches to complete a memory page.

Now the new I/O block layer enqueues a block request group with all the
buffers needed in a read or write operation, and keeps sleeping until the
whole transaction has been completed.

This improves dramatically the disk accesses, reduces a lot the number of
context switches, and it overall boosts the performance of the system.